rename some file
