+++ examples/llava_onevision/conf/compress/compress_llava_ov.yaml
@@ -1,21 +1,6 @@
-system:
-  tensor_model_parallel_size: 1
-  pipeline_model_parallel_size: 1
-  use_flash_attn: True
-  sequence_parallel: True
-  logging:
-    log_interval: 1
-    tensorboard_log_interval: 1
-    wandb_project: "compress-llavaov-7B" 
-    wandb_exp_name: "compress-test-7B" 
-  save_dir: 
-
-
-model: 
-  model_path: 
-  device_map: cuda:0
-  trust_remote_code: true
-  torch_dtype: bfloat16
+defaults:
+  - llava_ov_model
+  - _self_
 
 data:
   data_path: null
