+++ flagscale/compress/algo/algo_base.py
