+++ flagscale/compress/algo/smooth_quant.py
@@ -1,0 +1,7 @@
+from llmcompressor.modifiers.utils.compression_wrapper import ModuleCompressionWrapper
+from llmcompressor.modifiers.smoothquant import *
+
+class SmoothQuantWrapper(ModuleCompressionWrapper):
+    def __init__(self, name, layer):
+        super().__init__(name=name, layer=layer)
+
