+++ flagscale/compress/combined_algo.py
