+++ requirements/compress/requirements-dev.txt
@@ -1,0 +1,2 @@
+llmcompressor
+compressed-tensors-nightly
